feat: BROKEN - Partial implementation of bonsai update
The idea is to move from the genesets in a series of folders
to an adjacency matrix + list of nodes.

The thing is that we use `tree` to make the side of the 'pancan_heatmap'
and so it's hard to migrate until we find a substitute.

The implementation is only partial as this is a broken commit.