fix: fix NA/NULL mishap
The default value for `renames` was NULL, but it was interpreted by the
argument parsing library to be "there is no default", so it was falling
back to "NA" as the default. This broke the downsteam parsing if no
renames were supplied.

We now check for the NA and convert it to NULL on-demand.