fix: broken ties in the top_disregulation plot
This is done by summing ALL rank values (not just those of "significant"
tumor types) and sorting rows first by the number of tumors in which
that gene is "significantly different", and then by this total "rank
score".

I chose this (instead of only the "visible" signal) to account for all
the tumors that we cannot see, since it is a "shared" disregulation
plot, after all.